refactor: restructure Post module
Post now has two subpackages: api & impl. Code is currently in impl.
