+++ app/src/main/kotlin/com/potatosheep/kite/app/MainActivity.kt
@@ -33,7 +33,7 @@ import com.potatosheep.kite.app.ui.rememberAppState
 import com.potatosheep.kite.core.designsystem.KiteTheme
 import com.potatosheep.kite.feature.exception.ExceptionRoute
 import com.potatosheep.kite.feature.image.nav.navigateToImage
-import com.potatosheep.kite.feature.post.nav.navigateToPost
+import com.potatosheep.kite.feature.post.impl.nav.navigateToPost
 import com.potatosheep.kite.feature.subreddit.nav.navigateToSubreddit
 import dagger.hilt.android.AndroidEntryPoint
