+++ app/src/main/kotlin/com/potatosheep/kite/app/nav/KiteNavHost.kt
@@ -13,8 +13,8 @@ import com.potatosheep.kite.feature.bookmark.impl.nav.bookmarkScreen
 import com.potatosheep.kite.feature.image.nav.imageScreen
 import com.potatosheep.kite.feature.image.nav.navigateToImage
 import com.potatosheep.kite.feature.onboarding.impl.nav.onboardingScreen
-import com.potatosheep.kite.feature.post.nav.navigateToPost
-import com.potatosheep.kite.feature.post.nav.postScreen
+import com.potatosheep.kite.feature.post.impl.nav.navigateToPost
+import com.potatosheep.kite.feature.post.impl.nav.postScreen
 import com.potatosheep.kite.feature.search.nav.navigateToSearch
 import com.potatosheep.kite.feature.search.nav.searchScreen
 import com.potatosheep.kite.feature.settings.nav.settingsScreen
