+++ app/src/main/kotlin/com/potatosheep/kite/app/nav/TopLevelScreen.kt
@@ -60,7 +60,7 @@ import com.potatosheep.kite.feature.bookmark.impl.nav.navigateToBookmark
 import com.potatosheep.kite.feature.feed.impl.nav.FeedScreen
 import com.potatosheep.kite.feature.image.nav.navigateToImage
 import com.potatosheep.kite.feature.home.impl.nav.HomeScreen
-import com.potatosheep.kite.feature.post.nav.navigateToPost
+import com.potatosheep.kite.feature.post.impl.nav.navigateToPost
 import com.potatosheep.kite.feature.search.nav.navigateToSearch
 import com.potatosheep.kite.feature.settings.nav.navigateToSettings
 import com.potatosheep.kite.feature.subreddit.nav.navigateToSubreddit
