+++ feature/post/impl/src/main/kotlin/com/potatosheep/kite/feature/post/impl/PostScreen.kt
@@ -1,4 +1,4 @@
-package com.potatosheep.kite.feature.post
+package com.potatosheep.kite.feature.post.impl
 
 import androidx.compose.animation.AnimatedVisibility
 import androidx.compose.animation.fadeIn
