+++ feature/post/impl/src/main/res/mipmap-anydpi/ic_launcher_round.xml
