+++ settings.gradle.kts
@@ -35,7 +35,8 @@ include(":core:designsystem")
 include(":feature:subscription")
 include(":feature:home:impl")
 include(":feature:home:api")
-include(":feature:post")
+include(":feature:post:impl")
+include(":feature:post:api")
 include(":feature:image:impl")
 include(":feature:image:api")
 include(":feature:subreddit")
